Refactor GenericAccountingTariff to extend MeterTariff
